refactor: newItent 사용

+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/bottari/create/BottariCreateViewModel.kt
@@ -3,9 +3,9 @@ package com.bottari.presentation.view.home.bottari.create
 import androidx.lifecycle.ViewModel
 
 class BottariCreateViewModel : ViewModel() {
-    var bottariId: Int = 0
+    var bottariId: Long = 0
 
     fun createBottari(name: String) {
-        bottariId = 1
+        bottariId = 1L
     }
 }
